requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
